breakout FilenameMatcher and GetIdMatcher into their own files
