+++ src/main/components/matcher/FilegroupMatcher.cpp
@@ -1,45 +1,16 @@
 /*
- * VGMTrans (c) 2002-2014
+ * VGMTrans (c) 2002-2024
  * Licensed under the zlib license,
  * refer to the included LICENSE.txt file
- */
+*/
 
-#include "Matcher.h"
-#include "Root.h"
+#include "FilegroupMatcher.h"
+#include "Format.h"
+#include "VGMSeq.h"
+#include "VGMInstrSet.h"
+#include "VGMSampColl.h"
+#include "VGMColl.h"
 #include <regex>
-
-using namespace std;
-
-Matcher::Matcher(Format *format) {
-  fmt = format;
-}
-
-bool Matcher::onNewFile(std::variant<VGMSeq *, VGMInstrSet *, VGMSampColl *, VGMMiscFile *> file) {
-  if(auto seq = std::get_if<VGMSeq *>(&file)) {
-    onNewSeq(*seq);
-  } else if(auto instr = std::get_if<VGMInstrSet *>(&file)) {
-    onNewInstrSet(*instr);
-  } else if(auto sampcoll = std::get_if<VGMSampColl *>(&file)) {
-    onNewSampColl(*sampcoll);
-}
-
-  return false;
-}
-
-bool Matcher::onCloseFile(std::variant<VGMSeq *, VGMInstrSet *, VGMSampColl *, VGMMiscFile *> file) {
-  if(auto seq = std::get_if<VGMSeq *>(&file)) {
-    onCloseSeq(*seq);
-  } else if(auto instr = std::get_if<VGMInstrSet *>(&file)) {
-    onCloseInstrSet(*instr);
-  } else if(auto sampcoll = std::get_if<VGMSampColl *>(&file)) {
-    onCloseSampColl(*sampcoll);
-  }
-  return false;
-}
-
-// ****************
-// FilegroupMatcher
-// ****************
 
 FilegroupMatcher::FilegroupMatcher(Format *format) : Matcher(format) {}
