+++ src/main/formats/CapcomSnes/CapcomSnesFormat.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Format.h"
-#include "Matcher.h"
+#include "FilegroupMatcher.h"
 #include "CapcomSnesScanner.h"
