+++ src/main/formats/ChunSnes/ChunSnesFormat.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "Format.h"
-#include "Matcher.h"
+#include "FilegroupMatcher.h"
 #include "ChunSnesScanner.h"
