+++ src/main/formats/HOSA/HOSAFormat.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "HOSAScanner.h"
