+++ src/main/formats/ItikitiSnes/ItikitiSnesFormat.h
@@ -5,8 +5,7 @@
  */
 #pragma once
 #include "Format.h"
-#include "Matcher.h"
-#include "Root.h"
+#include "FilegroupMatcher.h"
 #include "ItikitiSnesScanner.h"
 
 constexpr uint8_t kItikitiSnesSeqNoteKeyBias = 24;
