+++ src/main/formats/NDS/NDSFormat.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "NDSScanner.h"
 #include "VGMColl.h"
