+++ src/main/formats/Org/OrgFormat.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "OrgScanner.h"
