+++ src/main/formats/PS1/PS1Format.h
@@ -5,9 +5,7 @@
  */
 #pragma once
 #include "Format.h"
-#include "Root.h"
-#include "Matcher.h"
-#include "VGMColl.h"
+#include "FilegroupMatcher.h"
 #include "PS1Seq.h"
 #include "formats/PS1/Vab.h"
 #include "PS1SeqScanner.h"
@@ -18,6 +16,5 @@
 
 BEGIN_FORMAT(PS1)
   USING_SCANNER(PS1SeqScanner)
-  //USING_MATCHER_WITH_ARG(SimpleMatcher, true)
   USING_MATCHER(FilegroupMatcher)
 END_FORMAT()
