+++ src/main/formats/SegSat/SegSatFormat.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "SegSatScanner.h"
