+++ src/main/formats/SquarePS2/SquarePS2Format.h
@@ -1,9 +1,7 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "SquarePS2Scanner.h"
-#include "Matcher.h"
-#include "VGMColl.h"
+#include "GetIdMatcher.h"
 
 // ***************
 // SquarePS2Format
@@ -12,7 +10,6 @@
 BEGIN_FORMAT(SquarePS2)
   USING_SCANNER(SquarePS2Scanner)
   USING_MATCHER(GetIdMatcher)
-  //USING_MATCHER(SimpleMatcher)
 END_FORMAT()
