+++ src/main/formats/TriAcePS1/TriAcePS1Format.h
@@ -1,6 +1,5 @@
 #pragma once
 #include "Format.h"
-#include "Root.h"
 #include "TriAcePS1Scanner.h"
 
 BEGIN_FORMAT(TriAcePS1)
