+++ src/ui/qt/SequencePlayer.cpp
@@ -8,7 +8,6 @@
 #include <algorithm>
 #include <cstddef>
 
-#include "Root.h"
 #include "bass.h"
 #include "bassmidi.h"
 #include "VGMColl.h"
